data conversion update

+++ lib/utils/chart_data.dart
@@ -2,6 +2,6 @@ class ChartData {
   const ChartData({this.x, this.time, this.weight});
 
   final int x;
-  final double time;
+  final int time;
   final double weight;
 }
